+++ README.md
@@ -22,7 +22,7 @@
 
 ## ✨ 주요 기능
 
-- **🌐 세련된 웹 UI**: **Next.js**와 **Material-UI(MUI)**를 사용하여 직관적이고 아름다운 다크 모드 인터페이스를 제공합니다.
+- **🌐 세련된 웹 UI**: **Next.js 15**와 **Material-UI(MUI)**를 사용한 컴포넌트 기반 아키텍처로 직관적이고 아름다운 다크 모드 인터페이스를 제공합니다.
 - **✍️ AI 기반 스타일 분석**: 단순 번역을 넘어, 소설의 첫 부분을 AI가 분석하여 **주인공의 이름**과 **핵심 서사 스타일**을 자동으로 정의합니다.
 - **🎨 사용자 제어 및 수정**: AI가 분석한 주인공 이름과 서사 스타일(서술 문체, 톤, 핵심 규칙 등)을 번역 시작 전에 사용자가 직접 확인하고 자유롭게 수정할 수 있습니다.
 - **📚 동적 문맥 관리**: 번역 과정에서 중요한 **용어(Glossary)**와 등장인물의 **말투(Character Style)**를 동적으로 구축하고 업데이트하여 일관성을 유지합니다.
@@ -53,12 +53,30 @@ Context-Aware-Translation/
 │   │   ├── app/                 # Next.js 앱 라우터 (페이지, 레이아웃)
 │   │   │   ├── (auth)/          # 인증 관련 페이지
 │   │   │   ├── community/       # 커뮤니티 관련 페이지
-│   │   │   ├── components/      # 공통 컴포넌트
+│   │   │   ├── components/      # 재사용 가능한 컴포넌트
+│   │   │   │   ├── ApiConfiguration/  # API 설정 컴포넌트
+│   │   │   │   ├── AdvancedSettings/  # 고급 번역 설정
+│   │   │   │   ├── FileUpload/        # 파일 업로드 처리
+│   │   │   │   ├── StyleConfiguration/# 스타일 및 용어집 설정
+│   │   │   │   ├── jobs/              # 번역 작업 관련 컴포넌트
+│   │   │   │   ├── sections/          # 페이지 섹션 컴포넌트
+│   │   │   │   └── translation/       # 번역 관련 컴포넌트
+│   │   │   ├── hooks/           # 커스텀 React 훅
+│   │   │   │   ├── useApiKey.ts         # API 키 관리
+│   │   │   │   ├── useTranslationJobs.ts# 번역 작업 상태 관리
+│   │   │   │   ├── useTranslationService.ts # 번역 서비스 로직
+│   │   │   │   └── useJobActions.ts     # 작업 액션 핸들러
+│   │   │   ├── types/           # TypeScript 타입 정의
+│   │   │   │   ├── job.ts              # 작업 타입
+│   │   │   │   └── translation.ts      # 번역 관련 타입
+│   │   │   ├── utils/           # 유틸리티 함수
+│   │   │   │   ├── api.ts              # API 통신 함수
+│   │   │   │   └── constants/          # 상수 정의
 │   │   │   ├── privacy/         # 개인정보 처리방침 페이지
 │   │   │   ├── favicon.ico      # 파비콘
 │   │   │   ├── globals.css      # 전역 CSS 스타일
 │   │   │   ├── layout.tsx       # 루트 레이아웃
-│   │   │   └── page.tsx         # 메인 페이지
+│   │   │   └── page.tsx         # 메인 페이지 (267 lines)
 │   │   └── theme.ts             # Material-UI 테마 설정
 │   └── public/                  # 정적 파일 디렉토리
 ├── core/                        # 🧠 핵심 번역 엔진
@@ -106,6 +124,27 @@ Context-Aware-Translation/
 ├── package.json                 # 📦 Node.js 의존성
 └── Dockerfile                   # 🐳 Docker 설정
 ```
+
+## 🏗️ 프론트엔드 아키텍처
+
+### 컴포넌트 구조
+- **모듈화된 컴포넌트**: 각 기능별로 독립적인 컴포넌트로 분리하여 재사용성과 유지보수성 향상
+- **커스텀 훅**: 비즈니스 로직을 `hooks/` 디렉토리의 커스텀 훅으로 분리하여 관심사 분리
+- **타입 안정성**: TypeScript를 활용한 강력한 타입 체크와 IntelliSense 지원
+- **최적화된 메인 페이지**: 776줄에서 267줄로 감소 (66% 코드 감소)
+
+### 주요 컴포넌트
+- `ApiConfiguration`: API 제공자 선택 및 키 관리
+- `FileUploadSection`: 파일 업로드 및 분석 처리
+- `TranslationSettings`: 고급 번역 설정 관리
+- `StyleConfigForm`: 스타일 및 용어집 구성
+- `JobsTable`: 번역 작업 목록 및 상태 관리
+
+### 커스텀 훅
+- `useTranslationService`: 번역 서비스 비즈니스 로직
+- `useJobActions`: 작업 관련 액션 (다운로드, 검증, 수정)
+- `useTranslationJobs`: 번역 작업 상태 관리
+- `useApiKey`: API 키 및 제공자 설정 관리
 
 ## 💬 커뮤니티 기능
